Add files via upload
FFR code will help in calculating the fractional flow reserve in a stenotic artery in CAD or in carotid artery. Factional Flow Reserve is a gold standard to measure the severity of the stenosis. Using the results of CFD ( ANSYS or OpenFOAM), FFR code will calculate the extent of the severity of the stenotic artery under the in-silico investigation.